create selectors for each store which are assigned the root state type
