+++ frontend/store/explore.store.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable no-param-reassign */
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { RootState } from '.';
 import { getExploreRandom } from '../services/apiService';
 
 export const getRandomMountains = createAsyncThunk('randomMountains/getRandomMountains', async () => {
@@ -13,10 +12,10 @@ interface RandomMountainsListState {
   loading: 'idle' | 'pending' | 'succeeded' | 'failed'
 }
 
-const initialRandomMountainsState: RandomMountainsListState = {
+const initialRandomMountainsState = {
   randomMountainsList: [],
   loading: 'idle',
-};
+} as RandomMountainsListState;
 
 const randomMountainsSlice = createSlice({
   name: 'randomMountains',
@@ -35,7 +34,5 @@ const randomMountainsSlice = createSlice({
     });
   },
 });
-
-export const randomMountainSelector = (state: RootState) => state.exploreRandomMountains;
 
 export default randomMountainsSlice.reducer;
