+++ frontend/store/login.store.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable no-param-reassign */
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { RootState } from '.';
 import { login } from '../services/apiService';
 
 export const postLogin = createAsyncThunk('login/postLogin', async (email: string) => {
@@ -10,14 +9,16 @@ export const postLogin = createAsyncThunk('login/postLogin', async (email: strin
 });
 
 interface LoginState {
-  userDetails: {}
+  userDetails: {
+    id: number
+  }
   loading: 'idle' | 'pending' | 'succeeded' | 'failed'
 }
 
-const initialLoginState: LoginState = {
+const initialLoginState = {
   userDetails: {},
   loading: 'idle',
-};
+} as LoginState;
 
 const loginSlice = createSlice({
   name: 'login',
@@ -36,7 +37,5 @@ const loginSlice = createSlice({
     });
   },
 });
-
-export const loginSelector = (state: RootState) => state.login;
 
 export default loginSlice.reducer;
